:sparkles: [FEAT] 시험 중 정보 Redis로 유지하기 #527

+++ src/main/java/swm_nm/morandi/domain/testStart/dto/TestCodeDto.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class TestCodeDto {
-    private Long problemNumber;
+    private Integer problemNumber;
     private String pythonCode;
     private String cppCode;
     private String javaCode;
